pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "scatter"
version = "0.1.0"
description = "Concurrent SSH executor for 100+ hosts (asyncio + AsyncSSH, Typer CLI, YAML inventory)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  { name = "Matt" }
]
dependencies = [
  "asyncssh>=2.14.1",
  "pyyaml>=6.0.2",
  "rich>=13.7.1",
  "typer>=0.12.3",
  "tenacity>=8.3.0",
  "uvloop>=0.20.0; platform_system == 'Linux'",
]

[project.optional-dependencies]
test = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
]

[project.scripts]
scatter = "scatter.cli:app"


